testing/web-platform/tests/workers/semantics/multiple-workers/003.html


















<!doctype html>
<title>creating 3 nested dedicated workers</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function() {
  var w1 = new Worker('#1');
  w1.onmessage = this.step_func(function(e) {
    assert_equals(e.data, '123');
    this.done();
  });
});
</script>
